requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pyTelegramBotAPI==4.21.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
